Rename Main to Debug
