+++ ui/src/main/java/com/ouchadam/loldr/debug/Presenter.java
@@ -1,4 +1,4 @@
-package com.ouchadam.loldr.main;
+package com.ouchadam.loldr.debug;
 
 import android.app.Activity;
 import android.view.View;
@@ -6,10 +6,10 @@ import android.widget.Toast;
 
 import com.ouchadam.loldr.ui.R;
 
-public class MainActivityPresenter {
+public class Presenter {
 
-    public static MainActivityPresenter onCreate(Activity activity, final Listener listener) {
-        activity.setContentView(R.layout.activity_main);
+    public static Presenter onCreate(Activity activity, final Listener listener) {
+        activity.setContentView(R.layout.activity_debug);
 
         activity.findViewById(R.id.button_anon_sign_in).setOnClickListener(new View.OnClickListener() {
             @Override
@@ -27,7 +27,7 @@ public class MainActivityPresenter {
 
         Toast.makeText(activity, "Hello world", Toast.LENGTH_SHORT).show();
 
-        return new MainActivityPresenter();
+        return new Presenter();
     }
 
     public interface Listener {
