+++ ui/src/main/res/layout/activity_debug.xml
