+++ ui/src/main/res/values/strings.xml
@@ -1,6 +1,6 @@
 <resources>
   <string name="app_name">lol;dr</string>
-  <string name="label_activity_main">lol;dr (debug)</string>
+  <string name="label_activity_debug">lol;dr (debug)</string>
   <string name="label_activity_post">lol;dr Post</string>
   <string name="label_activity_feed">lol;dr Feed</string>
   <string name="label_launcher_entry">@string/app_name</string>
